% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knit.R
\name{ap}
\alias{ap}
\title{Returns link to PDF if in HTML, otherwise return normal citation.}
\usage{
ap(citation)
}
\arguments{
\item{citation}{Citation string, so either `@a`, `[@a]` or `[@a; @b]`.}
}
\description{
Returns link to PDF if in HTML, otherwise return normal citation.
}
