% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{bracket}
\alias{bracket}
\title{Wrap A String in Brackets}
\usage{
bracket(x)
}
\arguments{
\item{x}{\verb{<character>} string}
}
\value{
\verb{<character>} string
}
\description{
Wrap A String in Brackets
}
\examples{
bracket("XYZ")

}
