% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/math.R
\name{calculate.ci.error}
\alias{calculate.ci.error}
\title{Returns error for confidence interval.}
\usage{
calculate.ci.error(n, s, p)
}
\arguments{
\item{n}{Sample size.}

\item{s}{Standard deviation.}

\item{p}{Significance level.}
}
\description{
Returns error for confidence interval.
}
