% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-frame.R
\name{change.column.type}
\alias{change.column.type}
\title{Change the column type for one or more columns.}
\usage{
change.column.type(df, cols, func)
}
\arguments{
\item{df}{Data frame.}

\item{cols}{Data frame columns.}

\item{func}{Function to apply to the columns.}
}
\description{
This function ignores columns specified in `cols` which are unavailable in the dataframe.
}
