% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{chop}
\alias{chop}
\title{Chop Vector by Group}
\usage{
chop(v, g)
}
\arguments{
\item{v}{vector}

\item{g}{group}
}
\value{
vector
}
\description{
Chop Vector by Group
}
\examples{
(v <- random_hcpcs(2))

(g <- sample(1:2, size = length(v), replace = TRUE))

chop(v, g)
}
