% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codex-package.R
\docType{package}
\name{codex-package}
\alias{codex}
\alias{codex-package}
\title{codex: What the Package Does (One Line, Title Case)}
\description{
What the package does (one paragraph).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://andrewallenbruce.github.io/codex/}
  \item \url{https://github.com/andrewallenbruce/codex}
  \item Report bugs at \url{https://github.com/andrewallenbruce/codex/issues}
}

}
\author{
\strong{Maintainer}: Andrew Bruce \email{andrewallenbruce@gmail.com} [copyright holder]

}
\keyword{internal}
