% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/math.R
\name{cohens.d}
\alias{cohens.d}
\title{Returns Cohen's d}
\usage{
cohens.d(n1, n2, m1, m2, s1, s2)
}
\arguments{
\item{n1}{Sample size for group 1.}

\item{n2}{Sample size for group 2.}

\item{m1}{Mean for group 1.}

\item{m2}{Mean for group 2.}

\item{s1}{Standard deviation for group 1.}

\item{s2}{Standard deviation for group 2.}
}
\description{
Returns Cohen's d
}
