% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/redundant.R
\name{contrast}
\alias{contrast}
\title{Contrast Redundant Characters}
\usage{
contrast(long, short)
}
\arguments{
\item{long}{\if{html}{\out{<character>}} vector}

\item{short}{\if{html}{\out{<character>}} vector}
}
\value{
\if{html}{\out{<character>}} vector
}
\description{
Contrast Redundant Characters
}
\examples{
contrast(
   long = c("39550", "92083", "C5274"),
   short = c("3", "5", "M", "U"))

}
