% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/versus_summaries.r
\name{counter}
\alias{counter}
\title{For a set of possible decks, order by probability of being best counter to given deck}
\usage{
counter(deck, vs_array, type = c("multi", "draft", "mono"))
}
\arguments{
\item{deck}{A character, giving the name of a single deck to counter.}

\item{vs_array}{The versus model array to use, as returned by
\code{\link{extract_vs_model_array}}.}

\item{type}{List of decks to use. \code{multi} uses all legal multicolour decks.
\code{draft} also includes decks where none of the specs are from the same
colour as the starter. \code{mono} only includes monocolour decks.}
}
\value{
A data.table with three columns: \code{deck} for the name of the
  countering deck, \code{prob_best} for the probability that it's the best counter
  deck, and \code{mean_win_prob} for its mean win probability against the deck.
}
\description{
For a set of possible decks, order by probability of being best counter to given deck
}
