% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{delist}
\alias{delist}
\title{Unlist with no names}
\usage{
delist(x)
}
\arguments{
\item{x}{A named or unnamed \verb{<list>}}
}
\value{
An unnamed \verb{<chr>} vector
}
\description{
Unlist with no names
}
\examples{
delist(list(x = "XYZ"))

delist(list("XYZ"))

}
