% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draft_strategy.r
\name{draft_move}
\alias{draft_move}
\title{Suggest picks for the Codex forum draft format}
\usage{
draft_move(
  vs_array,
  first_P2_move,
  first_P1_move,
  second_P2_move,
  second_P1_move,
  starters = codexdata::starters[base == "yes"],
  nicknames = codexdata::nicknames
)
}
\arguments{
\item{vs_array}{An 2-D array of opposed component pair effects, for a single
sample.}

\item{first_P2_move}{Either NULL, or a length-2 character vector, giving a
starter and a spec.}

\item{first_P1_move}{Either NULL, or a length-3 character vector, giving a
starter and two specs.}

\item{second_P2_move}{Either NULL, or a length-2 character vector, giving two
specs.}

\item{second_P1_move}{Either NULL, or a length-1 character vector, giving a
spec.}

\item{starters}{A data.table, containing \code{spec} and \code{starter}
columns. Defaults to the official starters and specs in
\code{codexdata::starters}.}

\item{nicknames}{A data.table containing \code{nickname} and \code{name}
columns. Defaults to the nicknames table in \code{codexdata}.}
}
\value{
A data.table, with at least one row for each possible next set of
  picks. Each row represents a possible ensuing series of picks to the end of
  selection, such that the ensuing picks are optimal for their respective
  player, in the minimax sense for win probability, assuming players are of
  equal skill. For example, if the next set of picks is the first set for
  player 2, then, for each possible such pick, there is a row for each
  following set of second picks, such that player two's second pick minimises
  the P1 win probability, bearing in mind possible ensuring second picks for
  player one, and player one's second pick maximises it. there is a row for
  each proceeding second pick for player 1 that maximises the win probability
  for player 1. This is done for each step in turn, so all picks after the
  first are optimal for their respective player. The table has a column for
  each remaining pick step, following by \code{P1} and \code{P2} columns for
  the resulting decks, and a \code{`P1 win probability`} column. The table is
  sorted in descending P1 win probability, then alphabetically by picks,
  starting with the first remaining step.
}
\description{
Given opposed component pair effects, determine the strength of possible sets
of picks for the next step in a forum draft.
}
\details{
The forum draft works as follows: Player two begins by choosing a starter and
a spec. Player one then chooses a starter and two specs. Player two chooses
two specs, then player one chooses their final spec.

Unlike other formats, players are not allowed to both choose the same starter
or spec, so a player's picks block their opponent's options. To prevent
someone being completely locked out of valid choices, decks are not required
to have any specs associated with the chosen starter.

A \code{NULL} value for a pick step indicates that the step has not already
occurred. The function decides on values for the missing steps.
}
