% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/versus_summaries.r
\name{DT_mean_matchup}
\alias{DT_mean_matchup}
\title{Create an HTML table widget for a mean matchup matrix}
\usage{
DT_mean_matchup(mean_matchup, prefix, ...)
}
\arguments{
\item{mean_matchup}{A matrix, containing a single-sample matchup matrix.}

\item{prefix}{A character, giving the prefix to use for the two deck/player
columns.}

\item{...}{Other arguments to datatable.}
}
\value{
A datatable.
}
\description{
Create an HTML table widget for a mean matchup matrix
}
