% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nash.r
\name{DT_nash}
\alias{DT_nash}
\title{Create an HTML table widget for a Nash equilibrium data.table}
\usage{
DT_nash(nash, round = 3, ...)
}
\arguments{
\item{nash}{A matrix, containing a single Nash equilibrium information table.}

\item{round}{An integer, giving the number of significant digits to round
deck weights to.}

\item{...}{Arguments to pass to \code{\link[DT]{datatable}}.}
}
\value{
A datatable.
}
\description{
Create an HTML table widget for a Nash equilibrium data.table
}
