% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarise_results.r
\name{extract_skill_ranks}
\alias{extract_skill_ranks}
\title{Extract rank distribution for player skills}
\usage{
extract_skill_ranks(x)
}
\arguments{
\item{x}{A data frame, giving sampled player skills in wide format.}
}
\value{
A tibble, giving sampled player skill ranks, with levels in order of average rank.
}
\description{
Extract rank distribution for player skills
}
