% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/versus_summaries.r
\name{extract_vs_model_array}
\alias{extract_vs_model_array}
\title{Extract versus model opposed component pair effect array}
\usage{
extract_vs_model_array(tidy_results)
}
\arguments{
\item{tidy_results}{A list, containing a versus model's tidy results.}
}
\value{
An array of opposed component pair effects, with dimensions equal to
  number of samples, then the number of starters plus numbers of specs twice.
}
\description{
Extract versus model opposed component pair effect array
}
