% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neo.R
\name{factor.neo}
\alias{factor.neo}
\title{Add factors to array of NEO domains; useful for sorting.}
\usage{
factor.neo(x)
}
\arguments{
\item{x}{A vector of distinct values.}
}
\description{
Allows Asterisk `*` as wildcard.
}
