% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vctrs.R
\name{gchop}
\alias{gchop}
\title{Chop Vector by Group}
\usage{
gchop(v, g)
}
\arguments{
\item{v}{\verb{<character>} vector}

\item{g}{\verb{<integer>} group}
}
\value{
\verb{<list>}
}
\description{
Chop Vector by Group
}
\examples{
(v <- random_hcpcs(2))

(g <- sample(1:2, size = length(v), replace = TRUE))

gchop(v, g)
}
