% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/versus_summaries.r
\name{get_matchup_array}
\alias{get_matchup_array}
\title{Create an array of matchup matrix samples for given decks/players}
\usage{
get_matchup_array(sim, deck_names, players = NULL, player_seed = 1)
}
\arguments{
\item{sim}{A list, containing a versus model's tidy results.}

\item{deck_names}{A character vector, giving standardised deck names to get
matchups for. Standardised names may used the package's default nicknames.}

\item{players}{A character vector, giving players to assign to the given
decks, in order. If NULL, players are not accounted for. Otherwise, the
length must be equal to the length of the deck names. Players not present
in the model information will have their skill samples generated.}

\item{player_seed}{An integer, giving the random seed used to generate
unknown players' skill samples.}
}
\value{
An array, with dimensions (sample, player 1, player 2).
}
\description{
Create an array of matchup matrix samples for given decks/players
}
