% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/versus_summaries.r
\name{get_matchups_from_array}
\alias{get_matchups_from_array}
\title{Get sample data.table for matchup samples array}
\usage{
get_matchups_from_array(matchup_array)
}
\arguments{
\item{matchup_array}{An array, giving samples of matchup matrix information.}
}
\value{
A data.table with four columns: \code{P1} and \code{P2} for the
  decks/players involved, \code{prob_matchup} for the sampled P1 win
  probability, and \code{prob_mean} for the mean P1 win probability over all
  samples for that matchup, pre-calculated for use in
  \code{\link{plot_matchup_samples}}.
}
\description{
Get sample data.table for matchup samples array
}
