% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nash.r
\name{get_nash_equilibria}
\alias{get_nash_equilibria}
\title{Get Nash equilibirum information for matchup matrix / samples array}
\usage{
get_nash_equilibria(matchups)
}
\arguments{
\item{matchups}{Either a matrix or an array, containing a Nash equilibirum
information.}
}
\value{
Either a matrix or an array of Nash equilibria, depending on the
  given input's class.
}
\description{
Get Nash equilibirum information for matchup matrix / samples array
}
