% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/versus_summaries.r
\name{get_player_skills}
\alias{get_player_skills}
\title{Get player skill sample array for given players, including unknown players}
\usage{
get_player_skills(sim, players, player_seed = 1)
}
\arguments{
\item{sim}{A list, containing a versus model's tidy results.}

\item{players}{A character vector, giving players to assign to the given
decks, in order. If NULL, players are not accounted for. Otherwise, the
length must be equal to the length of the deck names. Players not present
in the model information will have their skill samples generated.}

\item{player_seed}{An integer, giving the random seed used to generate
unknown players' skill samples.}
}
\value{
A matrix of log-odds player effects on P1 victory when the player is
  P1, with dimensions (sample, player).
}
\description{
Get player skill sample array for given players, including unknown players
}
