% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarise_results.r
\name{get_probability_largest}
\alias{get_probability_largest}
\title{Get probability of parameters being largest}
\usage{
get_probability_largest(samples)
}
\arguments{
\item{samples}{a tibble, containing a column for each parameter}
}
\value{
a data.table, with an \code{effect} grouping column and an
  \code{`effect size`} probability column
}
\description{
Get probability of parameters being largest
}
