% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_data_for_modelling.r
\name{get_tidy_model_results}
\alias{get_tidy_model_results}
\title{Get tidy model results from match data}
\usage{
get_tidy_model_results(
  results,
  model_data,
  matches,
  array = FALSE,
  log_lik = FALSE
)
}
\arguments{
\item{results}{A stan object.}

\item{model_data}{A list, containing data used for the simulation in
\code{results}.}

\item{matches}{A data.frame containing Codex match data.}

\item{array}{if TRUE, adds a \code{vs_array} element to the result,
containing the versus model array.}

\item{log_lik}{if TRUE, extracts log_lik samples to use in cross-validation.}
}
\value{
A list containing model data, tidy results, and simulation
  diagnostics.
}
\description{
Get tidy model results from match data
}
