% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/versus_summaries.r
\name{get_variances}
\alias{get_variances}
\title{Extract versus model component variance sample information}
\usage{
get_variances(sim, source = c("sd", "var"))
}
\arguments{
\item{sim}{A list, containing a versus model's tidy results.}

\item{source}{a character, indicating whether standard deviations or
variances are given in \code{sim}.}
}
\value{
A data.table.
}
\description{
Extract versus model component variance sample information
}
