% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{getem}
\alias{getem}
\title{Get named element from list}
\usage{
getem(ll, nm)
}
\arguments{
\item{ll}{named \verb{<list>}}

\item{nm}{\verb{<character>} element name}
}
\value{
named \verb{<list>} element
}
\description{
Get named element from list
}
\examples{
list(x1 = NA, x2 = "AA") |> getem("x2")

}
