% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grouping.R
\name{group_3}
\alias{group_3}
\title{Group 3}
\usage{
group_3(x)
}
\arguments{
\item{x}{\if{html}{\out{<list>}} of character vectors}
}
\value{
\if{html}{\out{<list>}} of character vectors
}
\description{
Group 3
}
\examples{
c("C75", "C97", "G02", "G04") |>
   split_lengths() |>
   remove_redundant() |>
   split_first() |>
   group_3()

}
