% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/groupers.R
\name{group_4}
\alias{group_4}
\title{Group Length 4}
\usage{
group_4(x)
}
\arguments{
\item{x}{\verb{<list>} of character vectors}
}
\value{
\verb{<list>} of character vectors
}
\description{
Group Length 4
}
\examples{
c("C75", "C97", "G02", "G04") |>
   split_lengths() |>
   remove_redundant() |>
   split_first() |>
   group_4()

}
