% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/groupers.R
\name{group_5}
\alias{group_5}
\title{Group Length 5}
\usage{
group_5(x)
}
\arguments{
\item{x}{\verb{<list>} of character vectors}
}
\value{
\verb{<list>} of character vectors
}
\description{
Group Length 5
}
\examples{
c("C75", "C97", "G02", "G04") |>
   split_lengths() |>
   remove_redundant() |>
   split_first() |>
   group_5()

}
