% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/groupers.R
\name{group_5_num}
\alias{group_5_num}
\title{Group Length 5 Numeric}
\usage{
group_5_num(x)
}
\arguments{
\item{x}{\verb{<list>} of character vectors}
}
\value{
\verb{<list>} of character vectors
}
\description{
Group Length 5 Numeric
}
\examples{
c("C7510", "C9710", "G0200", "G0400") |>
   split_lengths() |>
   remove_redundant() |>
   split_first() |>
   group_5()

}
