% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/approaches.R
\name{jaccard}
\alias{jaccard}
\title{Jaccard Similarity}
\usage{
jaccard(x, y)
}
\arguments{
\item{x, y}{vectors representing the two sets.}
}
\value{
\link{numeric} value
}
\description{
Calculate the Jaccard index between two sets.
}
\details{
The Jaccard index, also known as the Jaccard similarity coefficient, measures
the similarity between two sets by comparing the size of their intersection
to the size of their union.
}
\examples{
jaccard(c(1, 2, 3), c(2, 3, 4))

jaccard(factor(c(1, 2, 3)), factor(c(2, 3, 4)))

jaccard(c(1), c())

}
