% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{len_eq_one}
\alias{len_eq_one}
\title{Predicate Length Equals 1}
\usage{
len_eq_one(x)
}
\arguments{
\item{x}{\verb{<character>} string}
}
\value{
\verb{<logical>}
}
\description{
Predicate Length Equals 1
}
\examples{
len_eq_one(c("XYZ"))

}
