% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{len_gt_one}
\alias{len_gt_one}
\title{Predicate Length Greater Than 1}
\usage{
len_gt_one(x)
}
\arguments{
\item{x}{\verb{<character>} string}
}
\value{
\verb{<logical>}
}
\description{
Predicate Length Greater Than 1
}
\examples{
len_gt_one(c("XYZ"))

}
