% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/approaches.R
\name{lump}
\alias{lump}
\title{Lump Like Vectors Together}
\usage{
lump(x, threshold = 3)
}
\arguments{
\item{x}{\if{html}{\out{<numeric>}} vector}

\item{threshold}{\if{html}{\out{<numeric>}} threshold; default is \code{3}}
}
\value{
\if{html}{\out{<numeric>}} vector
}
\description{
Lump Like Vectors Together
}
\examples{
lump(c(1:3, 5, 8:10, 15))

}
