% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pins.R
\name{mount_board}
\alias{mount_board}
\title{Mount \link[pins:pins-package]{pins} board}
\usage{
mount_board(source = c("local", "remote"), package = "codex")
}
\arguments{
\item{source}{\verb{<character>}; \code{"local"} or \code{"remote"}}

\item{package}{\verb{<character>} package name}
}
\value{
\verb{<pins_board_folder>} or \verb{<pins_board_url>}
}
\description{
Mount \link[pins:pins-package]{pins} board
}
\keyword{internal}
