% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nash.r
\name{nash_mean}
\alias{nash_mean}
\title{Get mean Nash equilibrium over sample array}
\usage{
nash_mean(nash_array)
}
\arguments{
\item{nash_array}{An array, containing Nash samples.}
}
\value{
A data.table, giving the mean Nash equilibrium.
}
\description{
Get mean Nash equilibrium over sample array
}
