% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{parent}
\alias{parent}
\title{Wrap A String in Parentheses}
\usage{
parent(x)
}
\arguments{
\item{x}{\verb{<chr>} string}
}
\value{
\verb{<chr>} string
}
\description{
Wrap A String in Parentheses
}
\examples{
parent("XYZ")

}
