% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{parentheses}
\alias{parentheses}
\title{Wrap A String in Parentheses}
\usage{
parentheses(x)
}
\arguments{
\item{x}{\verb{<character>} string}
}
\value{
\verb{<character>} string
}
\description{
Wrap A String in Parentheses
}
\examples{
parentheses("XYZ")

}
