% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarise_results.r
\name{plot_flipped_effects_violin}
\alias{plot_flipped_effects_violin}
\title{Make horizontal violin plot with given titles}
\usage{
plot_flipped_effects_violin(
  effects,
  title,
  subtitle = NULL,
  key_name = effect,
  value_name = `effect size`
)
}
\arguments{
\item{effects}{a tibble, containing the columns to be plotted}

\item{title}{plot title}

\item{subtitle}{optional plot subtitle}

\item{key_name}{an unquoted expression, naming the grouping column}

\item{value_name}{an unquoted expression, naming the values column}
}
\value{
a ggplot object.
}
\description{
Make horizontal violin plot with given titles
}
