% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nash.r
\name{plot_matchup_samples}
\alias{plot_matchup_samples}
\title{Plot matchup sample data.table, using facets over match pairings}
\usage{
plot_matchup_samples(matchup_samples, break_interval = 1)
}
\arguments{
\item{matchup_samples}{A data.table, giving matchup sample information.}

\item{break_interval}{A number, giving the increment to use for the matchup axis.}
}
\value{
A ggplot object.
}
\description{
Plot matchup sample data.table, using facets over match pairings
}
