% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarise_results.r
\name{plot_player_skill}
\alias{plot_player_skill}
\title{Plot player skill distributions, sorted by mean skill}
\usage{
plot_player_skill(
  skills,
  turn = c("overall", "first", "second"),
  subtitle = "Mean-effects models"
)
}
\arguments{
\item{skills}{A data frame, containing player skill samples in tall format.}

\item{turn}{A character, describing the type of player skill.}

\item{subtitle}{A character, giving the plot subtitle. Set to `waiver()` to
have no subtitle}
}
\value{
A ggplot.
}
\description{
Plot player skill distributions, sorted by mean skill
}
