% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarise_results.r
\name{plot_probability_largest}
\alias{plot_probability_largest}
\title{Plot probabilities of parameters being largest, sorted by probability}
\usage{
plot_probability_largest(probabilities)
}
\arguments{
\item{probabilities}{a tibble, containing an \code{effect} grouping column
and an \code{`effect size`} probability column}
}
\value{
a ggplot object
}
\description{
Plot probabilities of parameters being largest, sorted by probability
}
