% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarise_results.r
\name{plot_rank}
\alias{plot_rank}
\title{Plot rank distribution for parameters}
\usage{
plot_rank(samples, group_name, model_name = ggplot2::waiver())
}
\arguments{
\item{samples}{a tibble, containing a column of sample values for each parameter}

\item{group_name}{a string, giving the name of the ranked group for use in the plot title}

\item{model_name}{a string, giving the name/description of the model used, to use as the plot subtitle}
}
\value{
a ggplot object
}
\description{
Plot rank distribution for parameters
}
