% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/versus_summaries.r
\name{plot_variances}
\alias{plot_variances}
\title{Plot versus model variance sample information}
\usage{
plot_variances(variances)
}
\arguments{
\item{variances}{A data.table of variances, as returned by
\code{\link{get_variances}}.}
}
\value{
A ggplot object.
}
\description{
Plot versus model variance sample information
}
