% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_data_for_modelling.r
\name{prepare_match_data_for_modelling}
\alias{prepare_match_data_for_modelling}
\title{Prepare match data for input to Stan model}
\usage{
prepare_match_data_for_modelling(matches, starters, nicknames, mean = FALSE)
}
\arguments{
\item{matches}{A data.frame containing Codex match data.}

\item{starters}{A data.frame containing starter/spec pairs.}

\item{nicknames}{A data.frame containing deck name/nickname pairs.}

\item{mean}{A boolean. If FALSE, prepares data for a simple non-opposed
model, i.e. where deck strength is independent of the opposing deck. If
TRUE, prepares data for an opposed model}
}
\value{
A list to pass to \code{stan()}'s \code{data} argument.
}
\description{
Prepare match data for input to Stan model
}
