% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nash.r
\name{print_nash}
\alias{print_nash}
\title{Print win probability information for matchup Nash equilibrium}
\usage{
print_nash(nash)
}
\arguments{
\item{nash}{A matrix, containing a single Nash equilibrium information table.}
}
\value{
A matrix, containing the win probability information for the Nash
  equilibrium.
}
\description{
Print win probability information for matchup Nash equilibrium
}
