% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/groupers.R
\name{process_groups}
\alias{process_groups}
\title{Process Groups}
\usage{
process_groups(x)
}
\arguments{
\item{x}{\verb{<list>} of character vectors}
}
\value{
\verb{<list>} of character vectors
}
\description{
Process Groups
}
\examples{
random_hcpcs(10) |>
   split_lengths() |>
   remove_redundant() |>
   split_first() |>
   process_groups()
}
