% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/random_hcpcs.R
\name{random_hcpcs}
\alias{random_hcpcs}
\title{Generate Random HCPCS Codes Vector}
\usage{
random_hcpcs(n = 10, lng = c(1:5))
}
\arguments{
\item{n}{\verb{<integer>} number of codes; default is \code{10}}

\item{lng}{\verb{<integer>} vector of lengths; default is \code{c(1:5)}}
}
\value{
\verb{<character>} vector
}
\description{
Generate Random HCPCS Codes Vector
}
\examples{
random_hcpcs()

random_hcpcs(50, 3)

}
