% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/random_hcpcs.R
\name{random_hcpcs2}
\alias{random_hcpcs2}
\title{Generate Random HCPCS Codes Vector}
\usage{
random_hcpcs2(n = 10)
}
\arguments{
\item{n}{\verb{<integer>} number of codes; default is \code{10}}
}
\value{
\verb{<character>} vector
}
\description{
Generate Random HCPCS Codes Vector
}
\examples{
random_hcpcs2()
}
\keyword{internal}
