% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/random_hcpcs.R
\name{random_hcpcs_vec}
\alias{random_hcpcs_vec}
\title{Generate Random HCPCS Codes Vector}
\usage{
random_hcpcs_vec(n = 10)
}
\arguments{
\item{n}{\if{html}{\out{<integer>}} number of codes}
}
\value{
\if{html}{\out{<character>}} vector
}
\description{
Generate Random HCPCS Codes Vector
}
\examples{
random_hcpcs_vec(n = 5)
}
