% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recombinators.R
\name{recombine_groups}
\alias{recombine_groups}
\title{Recombine Groups}
\usage{
recombine_groups(x)
}
\arguments{
\item{x}{\verb{<list>} of character vectors}
}
\value{
\verb{<list>} of character vectors
}
\description{
Recombine Groups
}
\examples{
random_hcpcs() |>
   split_lengths() |>
   remove_redundant() |>
   split_first() |>
   process_groups() |>
   reduce_groups() |>
   recombine_groups()
}
