% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reducers.R
\name{red1}
\alias{red1}
\title{Reduce Length 1}
\usage{
red1(x)
}
\arguments{
\item{x}{\verb{<list>} of character vectors}
}
\value{
\verb{<list>} of character vectors
}
\description{
Reduce Length 1
}
\examples{
random_hcpcs(20) |>
   split_lengths() |>
   split_first() |>
   process_groups() |>
   red1()

}
