% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reducers.R
\name{red2}
\alias{red2}
\title{Reduce Length 2}
\usage{
red2(x)
}
\arguments{
\item{x}{\verb{<list>} of character vectors}
}
\value{
\verb{<list>} of character vectors
}
\description{
Reduce Length 2
}
\examples{
x <- random_hcpcs(20) |>
   split_lengths() |>
   split_first() |>
   process_groups()

red2(x)

}
