% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reducers.R
\name{red3}
\alias{red3}
\title{Reduce Length 3}
\usage{
red3(x)
}
\arguments{
\item{x}{\verb{<list>} of character vectors}
}
\value{
\verb{<list>} of character vectors
}
\description{
Reduce Length 3
}
\examples{
random_hcpcs(20) |>
   split_lengths() |>
   split_first() |>
   process_groups() |>
   red3()

}
