% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reducers.R
\name{red4}
\alias{red4}
\title{Reduce Length 4}
\usage{
red4(x)
}
\arguments{
\item{x}{\verb{<list>} of character vectors}
}
\value{
\verb{<list>} of character vectors
}
\description{
Reduce Length 4
}
\examples{
random_hcpcs(20) |>
   split_lengths() |>
   split_first() |>
   process_groups() |>
   red4()

}
