% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reducers.R
\name{red5}
\alias{red5}
\title{Reduce Length 5}
\usage{
red5(x)
}
\arguments{
\item{x}{\verb{<list>} of character vectors}
}
\value{
\verb{<list>} of character vectors
}
\description{
Reduce Length 5
}
\examples{
random_hcpcs(20) |>
   split_lengths() |>
   split_first() |>
   process_groups() |>
   red5()

}
