% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reducers.R
\name{reduce_groups}
\alias{reduce_groups}
\title{Reduce Groups}
\usage{
reduce_groups(x)
}
\arguments{
\item{x}{\verb{<list>} of character vectors}
}
\value{
\verb{<list>} of character vectors
}
\description{
Reduce Groups
}
\examples{
random_hcpcs(20) |>
   split_lengths() |>
   split_first() |>
   process_groups() |>
   reduce_groups()

}
