% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reducers.R
\name{reduce_runs}
\alias{reduce_runs}
\title{Reduce Runs}
\usage{
reduce_runs(x)
}
\arguments{
\item{x}{\verb{<list>} of character vectors}
}
\value{
\verb{<list>} of character vectors
}
\description{
Reduce Runs
}
\examples{
x <- random_hcpcs(20) |>
   split_lengths() |>
   split_first() |>
   process_groups()

delist(x$g2) |>
   split_max_vlen() |>
   as.data.frame() |>
   purrr::map(uniq_narm) |>
   purrr::map(sort_order) |>
   purrr::map(reduce_runs)

}
