% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/versus_summaries.r
\name{reformat_mean_matchup}
\alias{reformat_mean_matchup}
\title{Reformat mean matchup array into tall data.table format}
\usage{
reformat_mean_matchup(mean_matchup, prefix)
}
\arguments{
\item{mean_matchup}{A matrix, containing a single-sample matchup matrix.}

\item{prefix}{A character, giving the prefix to use for the two deck/player
columns.}
}
\value{
A data.table.
}
\description{
Reformat mean matchup array into tall data.table format
}
