% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/math.R
\name{rn}
\alias{rn}
\title{Round a number to `n` digits and returns a string.}
\usage{
rn(x, n)
}
\arguments{
\item{x}{Number to be rounded.}

\item{n}{Number of digits after rounding.}
}
\description{
Round a number to `n` digits and returns a string.
}
