% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/redundant.R
\name{rr}
\alias{rr}
\title{Alias for \code{contrast()}}
\usage{
rr(...)
}
\arguments{
\item{...}{
  Arguments passed on to \code{\link[=contrast]{contrast}}
  \describe{
    \item{\code{long}}{\if{html}{\out{<character>}} vector}
    \item{\code{short}}{\if{html}{\out{<character>}} vector}
  }}
}
\description{
Alias for \code{contrast()}
}
