% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{sf_convert}
\alias{sf_convert}
\title{Convert string to stringfish}
\usage{
sf_convert(x)
}
\arguments{
\item{x}{\verb{<character>} vector}
}
\value{
\verb{<character>} stringfish vector
}
\description{
Convert string to stringfish
}
\examples{
sf_convert(random_hcpcs(500))

}
