% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{sf_detect}
\alias{sf_detect}
\title{Detect by Regex}
\usage{
sf_detect(s, p)
}
\arguments{
\item{s}{subject}

\item{p}{regex pattern}
}
\value{
\link{logical} vector
}
\description{
Detect by Regex
}
\examples{
sf_detect(random_hcpcs(), "[A-Z]{1}")

}
