% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{sf_extract}
\alias{sf_extract}
\title{Extract by Regex}
\usage{
sf_extract(s, p)
}
\arguments{
\item{s}{subject}

\item{p}{regex pattern}
}
\value{
vector
}
\description{
Extract by Regex
}
\examples{
sf_extract(random_hcpcs(), "[A-Z]{1}")

}
