% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{sf_remove}
\alias{sf_remove}
\title{Remove by Regex}
\usage{
sf_remove(s, p)
}
\arguments{
\item{s}{subject}

\item{p}{regex pattern}
}
\value{
vector
}
\description{
Remove by Regex
}
\examples{
sf_remove(LETTERS, "A")

sf_remove(paste0(LETTERS, collapse = ""), "A")

}
