% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{smush}
\alias{smush}
\title{Collapse a vector to length 1}
\usage{
smush(...)
}
\arguments{
\item{...}{A split \verb{<chr>} vector}
}
\value{
A collapsed \verb{<chr>} string
}
\description{
Collapse a vector to length 1
}
\examples{
smush(c("X", "Y", "Z"))

smush(random_hcpcs(2))

}
