% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{split_1}
\alias{split_1}
\title{Split Vector}
\usage{
split_1(x)
}
\arguments{
\item{x}{\verb{<character>} vector}
}
\value{
\verb{<character>} vector
}
\description{
Split Vector
}
\examples{
split_1(random_hcpcs(10))

}
