% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{split_at}
\alias{split_at}
\title{Split Vector by Index Subset}
\usage{
split_at(x, y = x, i, z)
}
\arguments{
\item{x}{\if{html}{\out{<character>}} vector}

\item{y}{\if{html}{\out{<character>}} vector}

\item{i}{\if{html}{\out{<integer>}} index start}

\item{z}{\if{html}{\out{<integer>}} index end}
}
\value{
\if{html}{\out{<character>}} vector
}
\description{
Split Vector by Index Subset
}
\examples{
x <- random_hcpcs(10)

end <- sf_extract(x$x5, "[A-Z]$")

beg <- smush(take_at(end, 5), sf_remove(end, "[A-Z]$"))

split_at(x = end, y = beg, i = 1, z = 2)

}
