% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/splitters.R
\name{split_end}
\alias{split_end}
\title{Split Character Vector at End}
\usage{
split_end(x)
}
\arguments{
\item{x}{\verb{<character>} vector}
}
\value{
\verb{<list>} of character vectors
}
\description{
Split Character Vector at End
}
\examples{
split_end(c("0741T", "E0628", "L4387", "0360T", "1127F", "0002M"))

}
