% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/splitters.R
\name{split_first}
\alias{split_first}
\title{Split Character Vector by Lengths}
\usage{
split_first(x)
}
\arguments{
\item{x}{\verb{<character>} vector}
}
\value{
\verb{<list>} of character vectors
}
\description{
Split Character Vector by Lengths
}
\examples{
random_hcpcs(5) |>
   split_lengths() |>
   remove_redundant() |>
   split_first()

}
