% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/splitters.R
\name{split_max_vlen}
\alias{split_max_vlen}
\title{Split String into Pieces}
\usage{
split_max_vlen(x)
}
\arguments{
\item{x}{\verb{<character>} vector}
}
\value{
\verb{<list>} of character vectors
}
\description{
Splits each string in a character vector into a fixed number of pieces,
returning a character matrix.
}
\examples{
fruits <- c("apples and oranges and pears and bananas", "pineapples and mangos and guavas")

split_max_vlen(fruits)

}
