% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{splits}
\alias{splits}
\title{Delist, Unname and Split a String}
\usage{
splits(x)
}
\arguments{
\item{x}{\verb{<chr>} string or named \verb{<list>} of \verb{<chr>} strings}
}
\value{
An unnamed \verb{<list>} of split \verb{<chr>} vectors
}
\description{
Delist, Unname and Split a String
}
\examples{
# unnamed vector
splits("XYZ")

# named vector
splits(c(x = "XYZ"))

# unnamed list with one element
splits(list("XYZ"))

# unnamed list with multiple elements
splits(list("YYY", "ZZZ"))

# named list with one element
splits(list(x = "XYZ"))

# named list with multiple elements
splits(list(x = "YYY", xx = "ZZZ"))

}
