% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{take_at}
\alias{take_at}
\title{Subset Vector at Index}
\usage{
take_at(x, i = 1)
}
\arguments{
\item{x}{\if{html}{\out{<character>}} vector}

\item{i}{\if{html}{\out{<integer>}} index}
}
\value{
\if{html}{\out{<character>}} vector
}
\description{
Subset Vector at Index
}
\examples{
take_at(random_hcpcs(2), 2)

}
