% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{unique_narm}
\alias{unique_narm}
\title{Unique Values with NAs Removed}
\usage{
unique_narm(x)
}
\arguments{
\item{x}{vector}
}
\value{
vector
}
\description{
Unique Values with NAs Removed
}
\keyword{internal}
