% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{unique_vlen}
\alias{unique_vlen}
\title{Unique Lengths of Vector}
\usage{
unique_vlen(x)
}
\arguments{
\item{x}{vector}
}
\value{
vector
}
\description{
Unique Lengths of Vector
}
\examples{
random_hcpcs(5) |> unique_vlen()

}
