% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codex-package.r
\docType{package}
\name{codex-package}
\alias{codex}
\alias{codex-package}
\title{Tidy Codex game data and summarise model results}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: Mark Webster \email{markwebster204@yahoo.co.uk}

}
\keyword{internal}
