% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{gchop}
\alias{gchop}
\title{Chop Vector by Group}
\usage{
gchop(v, g)
}
\arguments{
\item{v}{\if{html}{\out{<character>}} vector}

\item{g}{\if{html}{\out{<integer>}} group}
}
\value{
a \if{html}{\out{<list>}}
}
\description{
Chop Vector by Group
}
\examples{
(v <- random_hcpcs(2))

(g <- sample(1:2, size = length(v), replace = TRUE))

gchop(v, g)
}
