% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{letters_to_numbers}
\alias{letters_to_numbers}
\title{Convert Letters to Integers}
\usage{
letters_to_numbers(x)
}
\arguments{
\item{x}{vector of letters}
}
\value{
vector of integers
}
\description{
Convert Letters to Integers
}
\examples{
letters_to_numbers(LETTERS)
}
\keyword{internal}
