% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/approaches.R
\name{lump}
\alias{lump}
\title{Lump Like Vectors Together}
\usage{
lump(x, threshold = 3)
}
\arguments{
\item{x}{\link{numeric} vector}

\item{threshold}{\link{numeric} threshold; default is \code{3}}
}
\value{
\link{numeric} vector
}
\description{
Lump Like Vectors Together
}
\examples{
lump(1:3, 5:8)

}
