% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/random_hcpcs.R
\name{random_hcpcs}
\alias{random_hcpcs}
\title{Generate Random HCPCS Codes}
\usage{
random_hcpcs(n, l)
}
\arguments{
\item{n}{\if{html}{\out{<integer>}} number of codes}

\item{l}{\if{html}{\out{<integer>}} length of codes}
}
\value{
\if{html}{\out{<character>}} vector
}
\description{
Generate Random HCPCS Codes
}
\examples{
random_hcpcs(n = 5, l = 5)

}
