% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/redundant.R
\name{remove_redundant}
\alias{remove_redundant}
\title{Remove Redundant Characters}
\usage{
remove_redundant(x)
}
\arguments{
\item{x}{\verb{<list>} of vectors}
}
\value{
\verb{<list>} of vectors
}
\description{
Remove Redundant Characters
}
\examples{
random_hcpcs(15) |>
   split_lengths() |>
   remove_redundant()

}
