% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{replace_regex}
\alias{replace_regex}
\title{Replace Pattern Occurrences}
\usage{
replace_regex(x, p, r)
}
\arguments{
\item{x}{\verb{<character>} vector; strings to search in}

\item{p}{\verb{<character>} vector; search patterns}

\item{r}{\verb{<character>} vector with replacements for matched patterns}
}
\value{
\verb{<character>} vector
}
\description{
Replace Pattern Occurrences
}
\examples{
replace_regex(c('stringi R', 'REXAMINE', '123'), '( R|R.)', ' r ')

}
