% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/redundant.R
\name{rr}
\alias{rr}
\title{Alias for \code{contrast()}}
\usage{
rr(...)
}
\arguments{
\item{...}{arguments passed to \code{contrast()}}
}
\value{
\if{html}{\out{<character>}} vector
}
\description{
Alias for \code{contrast()}
}
\examples{
contrast(
   long = c("39550", "92083", "C5274"),
   short = c("3", "5", "M", "U"))

}
