% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{sf_detect}
\alias{sf_detect}
\title{Detect by Regex}
\usage{
sf_detect(s, p)
}
\arguments{
\item{s}{\verb{<character>} vector}

\item{p}{\verb{<character>} regex pattern}
}
\value{
\verb{<logical>} vector
}
\description{
Detect by Regex
}
\examples{
sf_detect(random_hcpcs(), "[A-Z]{1}")

}
