% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{sf_sub}
\alias{sf_sub}
\title{Subset Vector by Range}
\usage{
sf_sub(x, i = 1, z)
}
\arguments{
\item{x}{vector}

\item{i}{index start}

\item{z}{index end}
}
\value{
vector
}
\description{
Subset Vector by Range
}
\keyword{internal}
