% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{sort_order}
\alias{sort_order}
\title{Sort, Order and Collapse Vector}
\usage{
sort_order(x)
}
\arguments{
\item{x}{\verb{<character>} vector}
}
\value{
\verb{<character>} vector
}
\description{
Sort, Order and Collapse Vector
}
\examples{
(x <- sample(c(LETTERS, 0:9), size = 10))

sort_order(x)

}
