% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{split_at}
\alias{split_at}
\title{Split Vector by Index Subset}
\usage{
split_at(x, y = x, start, stop)
}
\arguments{
\item{x}{\verb{<character>} vector}

\item{y}{\verb{<character>} vector; default is \code{x}}

\item{start}{\verb{<integer>} index start}

\item{stop}{\verb{<integer>} index end}
}
\value{
\verb{<character>} vector
}
\description{
Split Vector by Index Subset
}
\examples{
x <- c("0741T", "E0628", "L4387", "0360T", "1127F", "0002M")

(end <- sf_extract(x, "[A-Z]$"))

(beg <- paste0(take_at(end, 5), sf_remove(end, "[A-Z]$")))

split_at(x = end, y = beg, start = 1, stop = 2)

}
