% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split_lengths.R
\name{split_end}
\alias{split_end}
\title{Split Character Vector at End}
\usage{
split_end(x)
}
\arguments{
\item{x}{\if{html}{\out{<character>}} vector}
}
\value{
\if{html}{\out{<list>}} of character vectors
}
\description{
Split Character Vector at End
}
\examples{
x <- random_hcpcs(10)

split_end(x$x5)

}
