% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split_lengths.R
\name{split_first}
\alias{split_first}
\title{Split Character Vector by Lengths}
\usage{
split_first(x, verbose = FALSE)
}
\arguments{
\item{x}{\if{html}{\out{<character>}} vector}

\item{verbose}{\if{html}{\out{<logical>}} print output}
}
\value{
\if{html}{\out{<list>}} of character vectors
}
\description{
Split Character Vector by Lengths
}
\examples{
random_hcpcs(5) |>
   split_lengths() |>
   remove_redundant() |>
   split_first()

}
