% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/splitters.R
\name{split_lengths}
\alias{split_lengths}
\title{Split Character Vector by Lengths}
\usage{
split_lengths(x)
}
\arguments{
\item{x}{\verb{<character>} vector}
}
\value{
\verb{<list>} of character vectors
}
\description{
Split Character Vector by Lengths
}
\examples{
random_hcpcs(5) |>
   split_lengths()

}
