% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{take_at}
\alias{take_at}
\title{Subset Vector at Index}
\usage{
take_at(x, i = 1)
}
\arguments{
\item{x}{vector}

\item{i}{index}
}
\value{
vector
}
\description{
Subset Vector at Index
}
\keyword{internal}
